Docker Compose: align with upstream WAII VISS Data Server Compose
Replace the placeholder service for the WAII VISS Data Server with the
Compose upstream in WAII at commit d9fea71 within docker-compose-rl.yml.
This should simplify future alignment with upstream.

To simplify bring up the RemotiveLabs feeder is removed, but we
would expect to add this feature once the core functionality is running.

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>