Docker Compose: fix WAII redis DB redis.conf configuration file
The upstream WAII Docker Compose uses a pre-built redis image but
provides a redis.conf configuration file. As there is no build step for
the associated redis service in the Compose we can not tell it where to
find the WAII source by using a build context as we do for the VISS
server.

Instead simply provide the relative path to the location of the file.

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>