WAII VISS: Add Apache IoTDB database to supported State Storage backends
SUMMARY
Currently WAII supports the SQLite and Redis databases as State Storage
backends. Extend WAII to support Apache IoTDB database as well.

As we are using a git submodule for the WAII VISS Data Server sources
and the changes are not yet upstreamed we must point to a fork that
carries the changes.

DETAIL
Both 'get' and 'set' are supported.

For development Apache IoTDB v1.2.2 and the Apache IoTDB Go client v1.1.7
were used.

Integration was sanity tested using the COVESA Central Data Service
Playground and the WAII HTML Client.

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>