WAII VISS: support Apache IoTDB config file
Update the WAII VISS Data Server submodule to add support for runtime
configuration for the Apache IoTDB connector via a config file.

The following notes are reproduced from the commit message:
The config file is named iotdb-config.json and placed in the vissv2server
directory. The file is in JSON format to follow the pattern set by
existing config files within the VISS server.

The contents of the file are not validated and are assumed to be both
complete and correct.

Example file:
{
		"host": "localhost",
		"port": "6667",
		"username": "root",
		"password": "root",
		"queryPrefixPath": "root.test2.dev1",
		"queryTimeout(ms)": 5000
}

host, port, username and password are the client configuration to be
used to create a session (connection) with the IoTDB Server.

queryPrefixPath is the IoTDB database timeseries prefix to be used to
access VSS nodes/keys in get/set operations.

queryTimeout(ms) is the timeout to be used in milliseconds when querying
the database.

If no config file is found on server startup the default (build-time)
configuration is used.

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>