VISS: take VISSR HEAD to integrate upstreamed Apache IoTDB support
Updates in VISS support:

1) UPSTREAM VISS PROJECT MOVE AND RENAME

The WAII data server project we use for VISS support has migrated
from W3C github infrastructure to COVESA github. At the same time it
was renamed from WAII to VISSR. Update the submodule to use the new name
and location.

2) UPDATE TO CURRENT UPSTREAM HEAD FOR APACHE IOTDB SUPPORT

Most importantly the changes to VISSR developed as part of this
project to add Apache IoTDB support as a data store backend have now
all been merged upstream. In summary these are in three areas: support
for IoTDB to get/set VSS data, document IoTDB in the VISSR documentation
site and provide runtime configuration of the connection parameters to
IoTDB. Rebasing on the current VISSR HEAD allows the playground to point
to the upstream rather than a fork carrying the changes.

Support was successfully sanity tested doing get/set operations using
Apache IoTDB and the VISSR html client.

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>