doc site: add Apache IoTDB and VISSR guides, also misc minor changes
1) Add a new "manuals" sub-tree as a first implementation of an "area"
to hold user guides, manuals, integration guides etc. Basically manuals
for the use of the playground.

2) Having done that populate it with guides for Apache IoTDB and VISSR
the two main components of the current docker deployment. The Apache
IoTDB guide is fairly comprehensive, whilst the VISSR guide is more of a
placeholder at this point. Update the main page to link to them both.

3) Take the opportunity to make a couple of minor fixes elsewhere.

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>