online doc: change IoTDB DB example to use a 'device'
Currently the example in IoTDB guide of the documentation site has a one
tier database/device path with the database root.test2.dev1 off which
the VSS measurements are added. This causes issues when you want to
select all data, e.g. 'select * from root.test2.dev1' will create an error.

It is best to have a 'device' separate from the 'database' in IoTDB
terms. So simply change the guide so that the 'create database' is for
'root.test2' only, which will solve both issues.

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>