remotivelabs example: add RL python subscribe sample as starting point
For the Spring 2024 COVESA AMM a simple bridge is desired to demonstrate
using the RemotiveLabs cloud as a feeder of vehicle data into the
playground data store.

As a starting point take the 'python subscribe' sample from the
RemotiveLabs sample repository [1]. We directly use the source rather
than a submodule to the complete repository for brevity and because we
expect the simple bridge will evolve into a more complete solution and
likely replaced.

[1] https://github.com/remotivelabs/remotivelabs-samples/tree/1939efca4ce5e10a7001595bb4cbe81de4a9a74e/python/subscribe

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>