remotivelabs example: implement simple Apache IoTDB bridge
Reimplement the simple bridge created for the Spring 2024 COVESA AMM.

The code is simpler if we do not need to query IoTDB to see what
timeseries schemas are in use. IoTDB can infer the data type from string
data for defined schemas. So use this ability to achieve the simplicity
goal. This also makes the example more readable.

The bridge waits for the subscribed signals to be sent from the
RemotiveLabs broker, formats them and then writes them to IoTDB using
the specified IoTDB path.

The code assumes the timeseries already exists.

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>