docker: add Dockerfile for CDSP Apache IoTDB image
Introduce a Dockerfile to build a CDSP specific Apache IoTDB image.

To keep the build as simple and quick as possible the starting point is
the pre-built IoTDB image from the upstream IoTDB project. This avoids
the burden of building IoTDB from scratch. This can be done because
currently our needs are limited to configuration and adding some
additional files.

Beyond some specific runtime environment values that can be hot loaded
IoTDB stores configuration in files located in the 'conf' directory.
For some specific use cases it is useful to automate the configuration
so end users of the Playground do not need to have to go through a lot
of manual configuration. For example, in this commit we enable the REST
API which is used by the IoTDB Grafana Connector to make DB queries.

In addition this commit integrates the optional IoTDB UDF (User Defined
Function) Data Quality Library into the image, which greatly extends
the data processing functions of IoTDB.

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>