doc site: documented integration of IoTDB Data Quality Library functions
The Playground docker deployment now includes the IoTDB project's
Data Quality Library of data processing functions in the Playground
IoTDB container image.

Rewrite the existing UDF section of the IoTDB reference manual to better
explain the built-in functions and those provided in the Data Quality
Library. Along with how to enable the Library functions.

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>