online doc: switch to git submodule for Hugo Learn theme. Fix page ToC
1) When the Hugo Learn theme was originally integrated into CDSP the
source was directly integrated into the CDSP source tree. To simplify
tracking the upstream git repository for the theme switch to using a
git submodule and update to the current version.

2) Updating to the current master branch fixes an issue in the theme
that resulted in level three headings on a page not being rendered so
they are indented correctly in the page pop-up Table of Contents.
This fixes CDSP bug #54 (online doc: Level 3 headings in generated ToC
are not indented correctly)

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>